+++ usr.bin/f77/libF77/r_lg10.c
@@ -3,12 +3,12 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_lg10.c	5.1	06/07/85
+ *	@(#)r_lg10.c	5.2	07/08/85
  */
 
 #define log10e 0.43429448190325182765
 
-double r_lg10(x)
+float r_lg10(x)
 float *x;
 {
 double log();
